tree: rework

+++ nvmecache/install.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env ash
+#!/usr/bin/env sh
 #
 # Copyright (C) 2025 AuxXxilium <https://github.com/AuxXxilium> and Ing <https://github.com/wjz304>
 #
@@ -6,51 +6,60 @@
 # See /LICENSE for more information.
 #
 
-check_model() {
-  MODELS="DS918+ RS1619xs+ DS1019+ DS718+ DS1621xs+"
-  MODEL="$(cat /proc/sys/kernel/syno_hw_version)"
+MODELS="DS918+ RS1619xs+ DS419+ DS1019+ DS719+ DS1621xs+"
+MODEL="$(cat /proc/sys/kernel/syno_hw_version)"
 
-  if ! echo "${MODELS}" | grep -wq "${MODEL}"; then
-    echo "${MODEL} is not supported nvmecache addon!"
-    exit 0
-  fi
-}
+if ! echo "${MODELS}" | grep -wq "${MODEL}"; then
+  echo "${MODEL} is not supported nvmecache addon!"
+  exit 0
+fi
 
-install_patches() {
-  echo "Installing addon nvmecache - patches"
+if [ "${1}" = "patches" ]; then
+  echo "Installing addon nvmecache - ${1}"
 
-  BOOTDISK_PART3_PATH="$(blkid -L ARC3 2>/dev/null)"
+  BOOTDISK_PART3_PATH="$(/sbin/blkid -L ARC3 2>/dev/null)"
   [ -n "${BOOTDISK_PART3_PATH}" ] && BOOTDISK_PART3_MAJORMINOR="$((0x$(stat -c '%t' "${BOOTDISK_PART3_PATH}"))):$((0x$(stat -c '%T' "${BOOTDISK_PART3_PATH}")))" || BOOTDISK_PART3_MAJORMINOR=""
-  [ -n "${BOOTDISK_PART3_MAJORMINOR}" ] && BOOTDISK_PART3="$(cat /sys/dev/block/${BOOTDISK_PART3_MAJORMINOR}/uevent 2>/dev/null | grep 'DEVNAME' | cut -d'=' -f2)" || BOOTDISK_PART3=""
+  [ -n "${BOOTDISK_PART3_MAJORMINOR}" ] && BOOTDISK_PART3="$(cat "/sys/dev/block/${BOOTDISK_PART3_MAJORMINOR}/uevent" 2>/dev/null | grep 'DEVNAME' | cut -d'=' -f2)" || BOOTDISK_PART3=""
 
   [ -n "${BOOTDISK_PART3}" ] && BOOTDISK="$(ls -d /sys/block/*/${BOOTDISK_PART3} 2>/dev/null | cut -d'/' -f4)" || BOOTDISK=""
-  [ -n "${BOOTDISK}" ] && BOOTDISK_PHYSDEVPATH="$(cat /sys/block/${BOOTDISK}/uevent 2>/dev/null | grep 'PHYSDEVPATH' | cut -d'=' -f2)" || BOOTDISK_PHYSDEVPATH=""
+  [ -n "${BOOTDISK}" ] && BOOTDISK_PHYSDEVPATH="$(cat "/sys/block/${BOOTDISK}/uevent" 2>/dev/null | grep 'PHYSDEVPATH' | cut -d'=' -f2)" || BOOTDISK_PHYSDEVPATH=""
 
   echo "BOOTDISK=${BOOTDISK}"
   echo "BOOTDISK_PHYSDEVPATH=${BOOTDISK_PHYSDEVPATH}"
 
   rm -f /etc/nvmePorts
-  for P in $(ls -d /sys/block/nvme* 2>/dev/null); do
-    if [ -n "${BOOTDISK_PHYSDEVPATH}" ] && [ "${BOOTDISK_PHYSDEVPATH}" = "$(cat ${P}/uevent 2>/dev/null | grep 'PHYSDEVPATH' | cut -d'=' -f2)" ]; then
-      echo "bootloader: ${P}"
+  for F in /sys/block/nvme*; do
+    [ ! -e "${F}" ] && continue
+    if [ -n "${BOOTDISK_PHYSDEVPATH}" ] && [ "${BOOTDISK_PHYSDEVPATH}" = "$(cat "${F}/uevent" 2>/dev/null | grep 'PHYSDEVPATH' | cut -d'=' -f2)" ]; then
+      echo "bootloader: ${F}"
       continue
     fi
-    PCIEPATH="$(cat ${P}/uevent 2>/dev/null | grep 'PHYSDEVPATH' | cut -d'=' -f2 | awk -F'/' '{if (NF == 4) print $NF; else if (NF > 4) print $(NF-1)}')"
+    PCIEPATH="$(cat "${F}/uevent" 2>/dev/null | grep 'PHYSDEVPATH' | cut -d'=' -f2 | awk -F'/' '{if (NF == 4) print $NF; else if (NF > 4) print $(NF-1)}')"
     if [ -n "${PCIEPATH}" ]; then
-      grep -q "${PCIEPATH}" /etc/nvmePorts && continue
-      echo "${PCIEPATH}" >> /etc/nvmePorts
+      grep -q "${PCIEPATH}" /etc/nvmePorts && continue # An nvme controller only recognizes one disk
+      echo "${PCIEPATH}" >>/etc/nvmePorts
     fi
   done
   [ -f /etc/nvmePorts ] && cat /etc/nvmePorts
-}
-
-install_late() {
-  echo "Installing addon nvmecache - late"
+elif [ "${1}" = "late" ]; then
+  echo "Installing addon nvmecache - ${1}"
   mkdir -p "/tmpRoot/usr/arc/addons/"
   cp -pf "${0}" "/tmpRoot/usr/arc/addons/"
 
+  #
+  # |       models      |     1st      |     2nd      |
+  # | DS918+            | 0000:00:13.1 | 0000:00:13.2 |
+  # | RS1619xs+         | 0000:00:03.2 | 0000:00:03.3 |
+  # | DS419+, DS1019+   | 0000:00:14.1 |              |
+  # | DS719+, DS1621xs+ | 0000:00:01.1 | 0000:00:01.0 |
+  #
+  # In the late stage, the /sys/ directory does not exist, and the device path cannot be obtained.
+  # (/dev/ does exist, but there is no useful information.)
+  # (The information obtained by lspci is incomplete and an error will be reported.)
+  # Therefore, the device path is obtained in the early stage and stored in /etc/nvmePorts.
+
   if [ ! -f /etc/nvmePorts ]; then
-    echo "/etc/nvmePorts does not exist"
+    echo "/etc/nvmePorts is not exist"
     exit 0
   fi
 
@@ -74,29 +83,9 @@ install_late() {
     fi
     idx=$((idx + 1))
   done
-}
-
-uninstall_addon() {
-  echo "Uninstalling addon nvmecache - ${1}"
+elif [ "${1}" = "uninstall" ]; then
+  echo "Installing addon nvmecache - ${1}"
 
   SO_FILE="/tmpRoot/usr/lib/libsynonvme.so.1"
   [ -f "${SO_FILE}.bak" ] && mv -f "${SO_FILE}.bak" "${SO_FILE}"
-}
-
-case "${1}" in
-  patches)
-    check_model
-    install_patches
-    ;;
-  late)
-    check_model
-    install_late
-    ;;
-  uninstall)
-    uninstall_addon
-    ;;
-  *)
-    exit 0
-    ;;
-esac
-exit 0
+fi
